+++ src/Globals.cpp
@@ -65,7 +65,7 @@ int ldrDivider = LDR_DIVIDER;
 const unsigned int LDR_RECOVER_TIME = 4000;
 unsigned long previousMillisLDR = 0;
 bool apFileRead;
-byte disconnectionCounter;
+int disconnectionCounter;
 
 /**
  * Set gpio received by the Firefly Luciferin software
